fix: rename file query_rewritter
